pyproject.toml
[tool.poetry]
name = "liga-magic"
version = "0.1.0"
description = ""
authors = ["Guilherme Noronha <guilhermenoronha2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.24.0"
webdriver-manager = "^4.0.2"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
